+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 # Change Log
 Notable changes will be documented here.
 
-## [TBD]
+## [0.24.0]
 - Skip keepAlive flag ([microsoft/vscode#228872](https://github.com/microsoft/vscode/issues/228872))
+- Refactor for reuse with fetch ([microsoft/vscode#228697](https://github.com/microsoft/vscode/issues/228697))
 
 ## [0.23.0]
 - Pass on keepAlive flag ([microsoft/vscode#173861](https://github.com/microsoft/vscode/issues/173861))